requirements.txt
streamlit
gspread
google-auth-oauthlib
google-auth-httplib2
requests